+++ src/polyfem/assembler/Assembler.hpp
@@ -36,11 +36,11 @@ namespace polyfem::assembler
 
 		virtual std::string name() const = 0;
 
-		int size() const { return size_; }
+		long int size() const { return size_; }
 		virtual void set_size(const int size) { size_ = size; }
 
 	protected:
-		int size_ = -1;
+		long int size_ = -1;
 
 		virtual int rows() const = 0;
 		virtual int cols() const = 0;
@@ -58,7 +58,7 @@ namespace polyfem::assembler
 
 		virtual std::string name() const = 0;
 
-		int size() const { return size_; }
+		long int size() const { return size_; }
 		virtual void set_size(const int size) { size_ = size; }
 
 		// assembler stiffness matrix, is the mesh is volumetric, number of bases and bases (FE and geom)
@@ -218,7 +218,7 @@ namespace polyfem::assembler
 		virtual bool is_tensor() const { return false; }
 
 	protected:
-		int size_ = -1;
+		long int size_ = -1;
 	};
 
 	// assemble matrix based on the local assembler
